+++ NEWS.md
@@ -15,6 +15,7 @@
 * New Emacs editing mode
 * More context-sensitive highlighting of R keywords
 * Option to enable highlighting of R function calls
+* F2 now navigates into files (e.g. within calls to source)
 
 ### R Markdown